requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
python-dotenv>=1.0.0
google-cloud-aiplatform>=1.56.0
google-auth>=2.29.0
google-api-python-client>=2.130.0
